Extract CqlQuery
